move compression utils out of HTTP into their own file
